botworld/botworld_api/pyproject.toml
[project]
name = "botworld"
version = "0.1.0"
description = ""
authors = [
    {name = "mirkuriit"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pygame (>=2.6.1,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "flask (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "flasgger (>=0.9.7.1,<0.10.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]
packages = [{include = "."}]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
